# docs/openapi/api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "后台管理接口",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/user/token": {
      "post": {
        "summary": "Rotate the dashboard PAT",
        "deprecated": false,
        "description": "Requires a live dashboard session and a one-time X-Security-Proof bound to scope pat.rotate and target self. Dashboard PAT authentication is rejected. The raw PAT is returned only once.",
        "tags": [
          "用户管理"
        ],
        "parameters": [
          {
            "name": "X-Security-Proof",
            "in": "header",
            "required": true,
            "description": "One-time proof for scope pat.rotate and target self",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PAT rotated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardAccessTokenRotationResponse"
                }
              }
            }
          },
          "401": {
            "description": "The dashboard Authorization credential is missing, expired, invalid, or no longer backed by an active login session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "403": {
            "description": "The one-time proof is missing, invalid, expired, consumed, or bound to a different scope or target",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "AccessToken1": [],
            "SecurityProof": []
          }
        ]
      },
      "delete": {
        "summary": "Revoke the dashboard PAT",
        "deprecated": false,
        "description": "Requires a live dashboard session and a one-time X-Security-Proof bound to scope pat.rotate and target self. Dashboard PAT authentication is rejected.",
        "tags": [
          "用户管理"
        ],
        "parameters": [
          {
            "name": "X-Security-Proof",
            "in": "header",
            "required": true,
            "description": "One-time proof for scope pat.rotate and target self",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PAT revoked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "The dashboard Authorization credential is missing, expired, invalid, or no longer backed by an active login session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "403": {
            "description": "The one-time proof is missing, invalid, expired, consumed, or bound to a different scope or target",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "AccessToken1": [],
            "SecurityProof": []
          }
        ]
      }
    },
...
    "/api/token/": {
      "get": {
        "summary": "获取所有令牌",
        "deprecated": false,
        "description": "🔐 需要登录（User权限）",
        "tags": [
          "令牌管理"
        ],
        "parameters": [
          {
            "name": "p",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      },
      "post": {
        "summary": "创建令牌",
        "deprecated": false,
        "description": "🔐 需要登录（User权限）",
        "tags": [
          "令牌管理"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Token"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      },
      "put": {
        "summary": "更新令牌",
        "deprecated": false,
        "description": "🔐 需要登录（User权限）",
        "tags": [
          "令牌管理"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Token"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
    "/api/token/{id}": {
      "get": {
        "summary": "获取指定令牌",
        "deprecated": false,
        "description": "🔐 需要登录（User权限）",
        "tags": [
          "令牌管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      },
      "delete": {
        "summary": "删除令牌",
        "deprecated": false,
        "description": "🔐 需要登录（User权限）",
        "tags": [
          "令牌管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
    "/api/usage/token/": {
      "get": {
        "summary": "获取令牌使用情况",
        "deprecated": false,
        "description": "🔑 需要令牌认证（TokenAuth）",
        "tags": [
          "令牌管理"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
    "/api/log/token": {
      "get": {
        "summary": "通过令牌获取日志",
        "deprecated": false,
        "description": "🔓 无需鉴权（通过令牌查询）",
        "tags": [
          "日志"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {}
        }
      },
      "DashboardAccessTokenRotation": {
        "type": "object",
        "required": [
          "token",
          "expires_at"
        ],
        "properties": {
          "token": {
            "type": "string",
            "pattern": "^pat_[0-9a-f]{64}$",
            "readOnly": true,
            "description": "The raw PAT. It is returned only in this rotation response and is never stored recoverably."
          },
          "expires_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp in seconds"
          }
        }
      },
      "DashboardAccessTokenRotationResponse": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/DashboardAccessTokenRotation"
          }
        }
      },
...
      "Token": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "expired_time": {
            "type": "integer"
          },
          "remain_quota": {
            "type": "integer"
          },
          "unlimited_quota": {
            "type": "boolean"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "SessionAuth1": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "已废弃：Gin session 不再用于面板鉴权"
      },
      "AccessToken1": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or PAT",
        "description": "面板短期 JWT 或 User.AccessToken PAT；不再要求 New-Api-User"
      },
      "NewApiUser1": {
        "type": "apiKey",
        "in": "header",
        "name": "New-Api-User",
        "description": "已废弃且不再参与鉴权；PAT 客户端可以移除此请求头"
      },
      "RefreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "new_api_refresh",
        "description": "仅 refresh/logout 使用的 HttpOnly、SameSite=Strict Cookie"
      },
      "SecurityProof": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Security-Proof",
        "description": "An exact-target, database-backed, one-time Security Proof bound server-side to the user, live browser session, user auth version, session version, method, scope, and target."
      },
      "Combination222": {
        "group": [
          {
            "id": 573667
          }
        ],
        "type": "combination"
      },
      "Combination1122": {
        "group": [
...
  },
  "servers": [],
  "security": [
    {
      "Combination343": []
    },
    {
      "Combination1243": []
    }
  ]
}